refactor: move billing and credit to lib/zero/
Move billing (3 source + 1 test) and credit (4 source + 2 test)
modules from src/lib/ to src/lib/zero/. These modules manage Zero
platform billing and credits, making them a natural fit for the
zero/ layer.

Closes #7730
