+++ turbo/apps/web/app/api/cron/cleanup-sandboxes/route.ts
@@ -19,7 +19,7 @@ import {
   drainOrgQueue,
   dispatchQueuedZeroRun,
 } from "../../../../src/lib/zero/zero-run-queue-service";
-import { processOrgCredits } from "../../../../src/lib/credit/credit-service";
+import { processOrgCredits } from "../../../../src/lib/zero/credit/credit-service";
 import { logger } from "../../../../src/lib/logger";
 import { env } from "../../../../src/env";
