+++ turbo/apps/web/app/api/cron/process-credits/route.ts
@@ -1,6 +1,6 @@
 import { NextResponse } from "next/server";
 import { initServices } from "../../../../src/lib/init-services";
-import { processStaleCredits } from "../../../../src/lib/credit/credit-service";
+import { processStaleCredits } from "../../../../src/lib/zero/credit/credit-service";
 import { logger } from "../../../../src/lib/logger";
 import { env } from "../../../../src/env";
