+++ turbo/apps/web/app/api/zero/billing/downgrade/route.ts
@@ -11,7 +11,7 @@ import {
   isAuthError,
 } from "../../../../../src/lib/auth/require-auth";
 import { resolveOrg } from "../../../../../src/lib/org/resolve-org";
-import { downgradeSubscription } from "../../../../../src/lib/billing/billing-service";
+import { downgradeSubscription } from "../../../../../src/lib/zero/billing/billing-service";
 
 const router = tsr.router(zeroBillingDowngradeContract, {
   create: async ({ body, headers }) => {
