+++ turbo/apps/web/app/api/zero/usage/members/route.ts
@@ -10,7 +10,7 @@ import {
   isAuthError,
 } from "../../../../../src/lib/auth/require-auth";
 import { resolveOrg } from "../../../../../src/lib/org/resolve-org";
-import { getUsageMembers } from "../../../../../src/lib/billing/usage-service";
+import { getUsageMembers } from "../../../../../src/lib/zero/billing/usage-service";
 
 const router = tsr.router(zeroUsageMembersContract, {
   get: async ({ headers }) => {
