+++ turbo/apps/web/src/lib/zero/billing/auto-recharge-service.ts
@@ -1,9 +1,9 @@
 import { eq, sql } from "drizzle-orm";
 import type Stripe from "stripe";
-import { orgMetadata } from "../../db/schema/org-metadata";
-import { grantOrgCredits } from "../org/org-service";
-import { getStripe } from "../stripe";
-import { logger } from "../logger";
+import { orgMetadata } from "../../../db/schema/org-metadata";
+import { grantOrgCredits } from "../../org/org-service";
+import { getStripe } from "../../stripe";
+import { logger } from "../../logger";
 
 const log = logger("billing:auto-recharge");
