+++ turbo/apps/web/src/lib/zero/billing/billing-service.ts
@@ -1,9 +1,9 @@
 import { eq } from "drizzle-orm";
 import type { OrgTier } from "@vm0/core";
-import { getStripe } from "../stripe";
-import { env } from "../../env";
-import { orgMetadata } from "../../db/schema/org-metadata";
-import { grantOrgCredits } from "../org/org-service";
+import { getStripe } from "../../stripe";
+import { env } from "../../../env";
+import { orgMetadata } from "../../../db/schema/org-metadata";
+import { grantOrgCredits } from "../../org/org-service";
 import { handleAutoRechargeInvoicePaid } from "./auto-recharge-service";
 import { resetMemberCreditFlags } from "../credit/member-credit-cap-service";
 import {
@@ -11,7 +11,7 @@ import {
   expireCredits,
   getExpiresRecordsSummary,
 } from "../credit/credit-expires-service";
-import { logger } from "../logger";
+import { logger } from "../../logger";
 
 const log = logger("billing");
