+++ turbo/apps/web/src/lib/zero/credit/__tests__/credit-expiry.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect, beforeEach } from "vitest";
-import { testContext, uniqueId } from "../../../__tests__/test-helpers";
+import { testContext, uniqueId } from "../../../../__tests__/test-helpers";
 import {
   insertCreditExpiresRecord,
   findCreditExpiresRecords,
@@ -7,7 +7,7 @@ import {
   grantCreditsToOrg,
   testDeductFromExpiresRecords,
   testExpireCredits,
-} from "../../../__tests__/api-test-helpers";
+} from "../../../../__tests__/api-test-helpers";
 import { getExpiresRecordsSummary } from "../credit-expires-service";
 
 const context = testContext();
