+++ turbo/apps/web/src/lib/zero/credit/credit-expires-service.ts
@@ -1,7 +1,7 @@
 import { eq, and, gt, lte, asc, sql } from "drizzle-orm";
-import { creditExpiresRecord } from "../../db/schema/credit-expires-record";
-import { orgMetadata } from "../../db/schema/org-metadata";
-import { logger } from "../logger";
+import { creditExpiresRecord } from "../../../db/schema/credit-expires-record";
+import { orgMetadata } from "../../../db/schema/org-metadata";
+import { logger } from "../../logger";
 
 const log = logger("service:credit-expires");
